+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Interprocess/Model/MessageCommandNames.cs
@@ -110,6 +110,7 @@
 
         // Layout, Gamepad
         public const string EnableGamepad = nameof(EnableGamepad);
+        public const string PreferDirectInputGamepad = nameof(PreferDirectInputGamepad);
 
         public const string GamepadHeight = nameof(GamepadHeight);
         public const string GamepadHorizontalScale = nameof(GamepadHorizontalScale);